+++ packages/gossip/test/coordinator/coordinator_lifecycle_test.dart
@@ -265,7 +265,7 @@ void main() {
         peerRepository: InMemoryPeerRepository(),
         entryRepository: InMemoryEntryRepository(),
         messagePort: port,
-        timerPort: InMemoryTimePort(),
+        timePort: InMemoryTimePort(),
       );
       addTearDown(coordinator.dispose);
       coordinator.errors.listen(errors.add);